Restructure database to allow deleting tasks via server documents.
took the task field from user, and put it into server, that means if I delete a server
document (for example the bot is kicked) all the tasks pertaining to the server will
be deleted along with it with no leftovers in the user document (apart from an element
in the servers array, which is deleted in a separate operation).

Need to update array on bot join, in case a user with a document is in the server (the
bot and the user have 2 common servers). Can't do this with fetch(), since that requires
privileged intent. I'll instead handle this during the registration check, by not only
checking for the existance of the user document, but also checking for the presence of
the user in the servers users array (or vice versa).