Add ability to catch up on launch
The bot can now catch up if it was removed from or added to a server whilst it was offline. This is predominantly implemented in the ready event handler itself. However, I also defined an object-making function that is now used both in the guildCreate handler, and the ready handler, to generate server objects to be inserted.